% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupar_partidos_uy.R
\name{agrupar_partidos_uy}
\alias{agrupar_partidos_uy}
\title{Agrupa partidos politicos}
\usage{
agrupar_partidos_uy(datos, umbral = 2)
}
\arguments{
\item{datos}{data.frame de clase \code{boreluy_elecciones} que devuelve la funcion \code{resultados_elecciones_uy}.}

\item{umbral}{Valor relativo al porcentaje bajo el cual se van a agrupar a los partidos como 'Otros Partidos'}
}
\value{
data.frame.
}
\description{
Permite agrupar partidos politicos de la salida de la funcion \code{resultado_eleccion_uy}.
    Junta las categorias de Voto en Blanco y Voto Anulado y en Otros Partidos a los que no alcanzan determinado umbral.
}
\examples{
elec99 <- resultado_eleccion_uy(anio = 1999, tipo = "Presidencial", por_departamento = FALSE)
agrupar_partidos_uy(elec99)
}
