% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_esaps.R
\name{as_esaps}
\alias{as_esaps}
\title{as_esaps}
\usage{
as_esaps(datos)
}
\arguments{
\item{datos}{Datos de la clase \code{boreluy}.}
}
\value{
data.frame.
}
\description{
Convierte salida de la funcion \code{nacional_uy} al formato necesario para
    poder usar el paquete \code{esaps}.
}
\examples{
elec <- nacional_uy(eleccion = 1971)

elec <- as_esaps(elec)

# Numero efectivo de partidos
# esaps::enp(elec)


# Desproporcionalidad electoral
# esaps::dispro(elec, method = 1)

# Nacionalizacion del sistema de partidos y de los partidos politicos
# esaps::psns(tidy_data = elec, method = 1, pns = TRUE)


}
