% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_parliament.R
\name{as_parliament}
\alias{as_parliament}
\title{as_parliament}
\usage{
as_parliament(datos, camara = 1, color = NULL, tipo = "semicircle", ...)
}
\arguments{
\item{datos}{Datos de la clase \code{boreluy_parlamento}.}

\item{camara}{Por defecto es 1 que correponde a la camara baja o Camara de Represnetantes, si se
indica 2 corresponde a la camara alta o Camara de Senadores.}

\item{color}{Paleta de colores con los que se van a identificar a los partidos politicos en el grafico.}

\item{tipo}{Tipo de parlamento (horseshoe, semicircle, circle, classroom, opposing benches)}

\item{...}{argumentos adicionales de la funcion \code{\link[ggparliament]{parliament_data}}}
}
\value{
data.frame.
}
\description{
Convierte salida de la funcion \code{\link{parlamento_uy}} al formato necesario para
    poder usar el paquete \code{ggparliament}.
}
