% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departamental_uy.R
\name{departamental_uy}
\alias{departamental_uy}
\title{Resultado de elecciones departamentales de Uruguay}
\usage{
departamental_uy(eleccion)
}
\arguments{
\item{eleccion}{Anio de eleccion.}
}
\value{
data.frame.
}
\description{
Devuelve el resultado electoral por eleccion departamentall en cantdiad de votos
    y en porcentaje.
}
