% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{departamentos}
\alias{departamentos}
\title{departamentos}
\format{
An object of class \code{character} of length 19.
}
\usage{
departamentos
}
\description{
contiene los nombres de los 19 departamentos de Uruguay
}
\keyword{datasets}
