% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elecciones.R
\name{elecciones}
\alias{elecciones}
\title{elecciones}
\usage{
elecciones(tipo = character())
}
\arguments{
\item{tipo}{tip de eleccion}
}
\description{
Permite checar las elecciones disponibles en la base de datos \code{elecciones_uy}.
}
\examples{
elecciones(tipo = "Presidencial")
}
