% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elecciones_uy.R
\docType{data}
\name{elecciones_uy}
\alias{elecciones_uy}
\title{elecciones_uy}
\format{data.frame con 18122 filas y 24 columnas.}
\usage{
data(elecciones_uy)
}
\description{
Contiene el resultado de un conjunto de elecciones de nivel nacional de Uruaguat entre 1918 y 2020.
}
\details{
\describe{
   \item{anio_eleccion}{Anio en el que se realizo la eleccion.}
   \item{fecha}{Fecha en la que se realizo la eleccion.}
   \item{eleccion}{Tipo de eleccion.}
   \item{concurrente}{Si hay elecciones concurrentes. Si la hay vale 1, en caso contrario la variable asume el }
   \item{parcial}{...}
   \item{organo}{...}
   \item{cargos}{Cantidad principal de cargos que se eligen en cada eleccion.}
   \item{partido}{Nombre del partido politico.}
   \item{departamento}{Nombre del departamento.}
   \item{fraccion}{Nombre de la fraccion del partido politico.}
   \item{sublema_senado}{Nombre del sublema al senado.}
   \item{cabeza_senado}{Nombre de la persona que encabeza la lista al senado.}
   \item{votos}{Cantidad de votos.}
   \item{bancas_diputados}{Cantidad de bancas a la Camara de Representantes obtenidas por el partido politico.}
   \item{bancas_senado}{Cantidad de bancas a la Camara de Senadores obtenidas por el partido politico.}
   \item{total_bancas_diputados}{Total de bancas que tiene la Camara de Representantes.}
   \item{total_bancas_senado}{Total de bancas que tiene la Camara de Senadores.}
   \item{habilitados}{Cantidad de ciudadanos habilitados para votar.}
   \item{poblacion}{Cantidad de poblacion.}
   \item{formula_presidente}{Candidato a la presidncia.}
   \item{formula_vice}{Candidato a la vicepresidencia.}
   \item{partido_ganador}{Partido politico ganador de la eleccion.}
   \item{presidente}{Presidente electo.}
   \item{vice}{Vicepresidente electo.}
}

La unidad de analisis de esta base de datos es la fraccion a nivel departamental. Esto hace que se tenga el resultado electoral estructurado por cada fraccion de cada partido en cada departamento.
}
\keyword{datasets}
