% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen_elecciones_uy.R
\name{grafico_elecciones_uy}
\alias{grafico_elecciones_uy}
\title{grafico elecciones uy}
\usage{
grafico_elecciones_uy()
}
\description{
Resumen de la cantidad de elecciones por organo y por anio de forma grafica.
}
