% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacional_uy.R
\name{nacional_uy}
\alias{nacional_uy}
\title{Resultado de elecciones nacionales de Uruguay}
\usage{
nacional_uy(
  eleccion,
  tipo = "Presidencial",
  por_departamento = FALSE,
  vbva.rm = FALSE
)
}
\arguments{
\item{eleccion}{Anio de eleccion.}

\item{tipo}{Tipo de eleccion. POr defecto es 'Presidencial'}

\item{por_departamento}{Por defecto es \code{FALSE}. Si es \code{TRUE} devuelve el resultado por departamento.}

\item{vbva.rm}{Eliminar del calculo a los votos en blanco y anulados. Esto puede ser util para elecciones como el 'Balotaje'.}
}
\value{
data.frame. Si la eleccion es concurrente (hay eleccion de cargos legislativos)
    devuelve la cantidad de bancas por partido. Si la eleccion no es concurrente solo devuelve
    el resultado en porcentaje de la eleccion por partido.
}
\description{
Devuelve el resultado electoral por eleccion nacional en cantdiad de votos
    y en porcentaje. Las eleccciones nacionales son: Presidencial, Balotaje, Legislativa y del
    Consejo Nacional de Administracion.
}
\examples{
nacional_uy(eleccion = 1971)
nacional_uy(eleccion = 1999, por_departamento = TRUE)
nacional_uy(eleccion = 1999, tipo = 'Balotaje')
nacional_uy(1925, tipo = 'Legislativa')
}
