% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados_elecciones_uy.R
\name{resultado_eleccion_uy}
\alias{resultado_eleccion_uy}
\title{Resultados electorales de elecciones nacionales y departamemntales}
\usage{
resultado_eleccion_uy(
  anio = integer(),
  tipo = NULL,
  por_departamento = FALSE,
  vbva.rm = FALSE
)
}
\arguments{
\item{anio}{Anio de eleccion.}

\item{tipo}{Tipo de eleccion.}

\item{por_departamento}{Por defecto es \code{FALSE}. Si es \code{TRUE} devuelve el resultado por departamento.}

\item{vbva.rm}{Eliminar del calculo a los votos en blanco y anulados. Esto puede ser util para elecciones como el 'Balotaje'.}
}
\value{
data.frame.
}
\description{
Devuelve el resultado electoral por eleccion en cantdiad de votos
    y en porcentaje.
}
\examples{
resultado_eleccion_uy(anio = 1999, tipo = "Presidencial", por_departamento = FALSE)
}
