% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultados_elecciones_uy.R
\name{resultado_elecciones_uy}
\alias{resultado_elecciones_uy}
\title{Resultados electorales de elecciones nacionales y departamemntales}
\usage{
resultado_elecciones_uy(
  anio = max(elecciones_uy$anio),
  tipo = NULL,
  por_departamento = FALSE
)
}
\arguments{
\item{anio}{Anio de eleccion.}

\item{tipo}{Tipo de eleccion.}

\item{por_departamento}{Por defecto es \code{FALSE}. Si es \code{TRUE} devuelve el resultado por departamento.}
}
\value{
data.frame.
}
\description{
Devuelve el resultado electoral por eleccion en cantdiad de votos
    y en porcentaje.
}
\examples{
resultado_elecciones_uy(anio = 1999, tipo = "Presidencial", por_departamento = FALSE)
}
