% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultado_parlamento_uy.R
\name{resultado_parlamento_uy}
\alias{resultado_parlamento_uy}
\title{Composicion del parlamento}
\usage{
resultado_parlamento_uy(anio = 1971, por_departamento = FALSE)
}
\arguments{
\item{anio}{Anio de eleccion}

\item{por_departamento}{Por defecto es \code{FALSE}. Si es \code{TRUE} devuelve
el resultado por departamento, como la camara alta es de circunscrupcion nacional
el resultado va a ser contsnte en todos los departamemtos.}
}
\value{
data.frame.
}
\description{
Devuelve la integracin del parlameto (camara baja y camara alta)
    en cantidad de bancas por partido y por departamento.
}
\examples{
resultado_parlamento_uy(1971, FALSE)
}
