% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumen_elecciones_uy.R
\name{resumen_elecciones_uy}
\alias{resumen_elecciones_uy}
\title{resumen elecciones uy}
\usage{
resumen_elecciones_uy(tabla = TRUE)
}
\arguments{
\item{tabla}{Por defecto es  \code{TRUE}. Devuelve una tabla resumen por anio.}
}
\value{
Si tabla es \code{TRUE} devuelve un data.frame, si es  \code{FALSE} imprime un resumen.
}
\description{
Resumen de la cantidad de elecciones por organo y por anio.
}
\examples{
resumen_elecciones_uy()
}
