% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupar_partidos_uy.R
\name{agrupar_partidos_uy}
\alias{agrupar_partidos_uy}
\title{Agrupa partidos politicos}
\usage{
agrupar_partidos_uy(datos, umbral = 2)
}
\arguments{
\item{datos}{data.frame que devuelve la funcion \code{nacional_uy} y \code{departamental_uy}.}

\item{umbral}{Valor relativo al porcentaje bajo el cual se van a agrupar a los partidos como 'Otros Partidos'}
}
\value{
data.frame.
}
\description{
Permite agrupar partidos politicos de la salida de la funcion \code{nacional_uy} y \code{departamental_uy}.
    Junta las categorias de Voto en Blanco y Voto Anulado y en Otros Partidos a los que no alcanzan determinado umbral.
}
\examples{
elec71 <- nacional_uy(eleccion = 1971)
agrupar_partidos_uy(elec71)
}
