% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_esaps.R
\name{as_esaps}
\alias{as_esaps}
\title{as_esaps}
\usage{
as_esaps(datos)
}
\arguments{
\item{datos}{Datos de la clase \code{boreluy_elecciones}.}
}
\value{
data.frame.
}
\description{
Convierte salida de la funcion \code{\link{resultados_elecciones_uy}} al formato necesario para
    poder usar el paquete \code{esaps}.
}
