% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_parliament.R
\name{as_parliament}
\alias{as_parliament}
\title{as_parliament}
\usage{
as_parliament(datos, camara = 1, color = NULL, tipo = "semicircle", ...)
}
\arguments{
\item{datos}{Datos de la clase \code{boreluy_parlamento}.}

\item{camara}{Por defecto es 1 que correponde a la camara baja o Camara de Represnetantes, si se
indica 2 corresponde a la camara alta o Camara de Senadores.}

\item{color}{Paleta de colores con los que se van a identificar a los partidos politicos en el grafico.}

\item{tipo}{Tipo de parlamento (horseshoe, semicircle, circle, classroom, opposing benches)}

\item{...}{argumentos adicionales de la funcion \code{\link[ggparliament]{parliament_data}}}
}
\value{
data.frame.
}
\description{
Convierte salida de la funcion \code{\link{resultado_parlamento_uy}} al formato necesario para
    poder usar el paquete \code{ggparliament}.
}
\examples{
diputados <- resultado_parlamento_uy(anio = 1971, por_departamento = FALSE)
as_parliament(diputados, camara = 1, color = c('#E81B23', '#3333FF', '#B4B4B4'))

# ggplot2::ggplot(diputados, aes(x, y, colour = party_long)) +
#    ggplot2::geom_parliament_seats() +
#    ggplot2::geom_highlight_government(government == 1) +
#    ggplot2::draw_majoritythreshold(n = mayoria, label = FALSE, type = 'semicircle') +
#    ggplot2::geom_parliament_bar(colour = colour, party = party_long) +
#    ggplot2::theme_ggparliament() +
#    ggplot2::labs(colour = NULL, title = titulo, subtitle = "") +
#    ggplot2::scale_colour_manual(values = datos$colour, limits = datos$party_long)

}
